+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Versions follow [CalVer](https://calver.org).
 
 ### Added
 
+- Define new methods to transform a `gspits.system.Partition` instance:
+  `with_size` (class method), `make_origin_centered_unit` (class method),
+  `origin_centered_unit`, `origin_centered`, `scaled`, and `translated`.
 - Add protocol `gspits.system.ExternalPotential` to define an external
   potential.
 - Add `gspits.system` module to implement protocols and classes used to define